requirements.txt
alembic==1.4.3
aniso8601==8.1.0
appdirs==1.4.4
attrs==20.3.0
black==20.8b1
cffi==1.14.4
click==7.1.2
coverage==5.3.1
cryptography==3.3.1
flake8==3.8.4
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
regex==2020.11.13
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.21
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
watchdog==1.0.1
Werkzeug==1.0.1
